Global Var vs Local Var
I spent a lot of time learning the difference
between global and local variables.

after that I started to understand that, I was able to
make the '+' on my calculator work.

I am pretty pumped about it. But i need to go to work
now, so I have to put the rest on hold.